week03-homework/pyproject.toml
[project]
name = "week03-homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"python-dotenv",
	"llama-index",
	"llama-index-llms-openai",
	"llama-index-embeddings-openai",
	"charset-normalizer",
	"matplotlib",
	"paddlepaddle>=3.0,<3.3",
	"paddleocr>=3.0,<3.1",
	"rapidocr-onnxruntime",
	# add your dependencies here and run `uv sync` to install them
]
